Fixed bug #1: Queue now works standard compilant
Before there was no way to have the same song twice in the Queue.
This was because pos/idx was saved in the songs table, forcing a 1:1 relation.
Now queue was split into a new table, with the sideeffect of even faster/less
hacky build-up! (because of most of the songs_content/c0uri stuff was removed.)