Added support of metadata-threads to libmoosecat
Metadata-threads provide true c-threads for paralell fetching.
Current solution is based on python threads, which aren't a nice thing, since it
involves a lot of context switches in the interpreter with many threads.

Will replace those in metadata.py soon.
The trick we gonna use: query.commit() is implemented to use no gil since it'a s
C-only function.